requirements.txt
click==8.3.1
colorama==0.4.6
joblib==1.5.3
nltk==3.9.4
numpy==2.4.4
pandas==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.3.32
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.3
tzdata==2025.3
